refactor: move counting iterator to pkg/iterator
Move countingIterator from pkg/blobby to pkg/iterator package where it
belongs alongside other iterator utilities. Rename to follow Go naming
conventions with exported type and constructor.

- Move to pkg/iterator/counting.go
- Rename countingIterator -> Counting
- Rename newCountingIterator -> NewCounting
- Update import and usage in archive.go
